variable change to handle larger numbers

+++ 0x04-more_functions_nested_loops/101-print_number.c
@@ -8,7 +8,7 @@
  */
 void print_number(int n)
 {
-	int l, c = 1;
+	long int l, c = 1;
 
 	if (n < 0)
 	{
